fix: move entity from types.go to entity.go
